requirements.txt
apturl==0.5.2
asgiref==3.4.1
attrs==19.3.0
bcrypt==3.1.7
blinker==1.4
boto3==1.20.12
botocore==1.23.12
Brlapi==0.7.0
bunnycdnpython==0.0.7
cached-property==1.5.1
cachetools==5.0.0
certifi==2021.10.8
cffi==1.14.6
chardet==3.0.4
charset-normalizer==2.0.6
chrome-gnome-shell==0.0.0
Click==7.0
colorama==0.4.4
command-not-found==0.3
cryptography==35.0.0
cupshelpers==1.0
dbus-python==1.2.16
decorator==4.4.2
defer==1.0.6
defusedxml==0.7.1
distlib==0.3.4
distro==1.4.0
distro-info===0.23ubuntu1
dj-database-url==0.5.0
Django==3.2.8
django-admin-honeypot==1.1.0
django-allauth==0.50.0
django-bootstrap-datepicker-plus==3.0.5
django-ckeditor==6.1.0
django-cleanup==5.2.0
django-crispy-forms==1.13.0
django-crum==0.7.9
django-debug-toolbar==3.2.2
django-embed-video==1.4.0
django-hosts==5.1
django-js-asset==1.2.2
django-modeltranslation==0.17.3
django-multiselectfield==0.1.12
django-next-prev==1.1.0
django-phonenumber-field==6.1.0
django-protected-media==1.0.0
django-recaptcha==2.0.6
django-recurrence==1.10.3
django-rosetta==0.9.7
django-storages==1.12.3
django-taggit==1.5.1
django-tempus-dominus==5.1.2.16
docker==4.1.0
docker-compose==1.25.0
dockerpty==0.4.1
docopt==0.6.2
duplicity==0.8.12.0
entrypoints==0.3
ez-setup==0.9
fasteners==0.14.1
filelock==3.6.0
future==0.18.2
GDAL==3.0.4
google-auth==2.6.2
gunicorn==20.1.0
httplib2==0.14.0
icalendar==4.0.9
idna==3.2
imageio==2.9.0
imageio-ffmpeg==0.4.5
importlib-metadata==1.5.0
jmespath==0.10.0
jsonschema==3.2.0
keyring==18.0.1
keyrings.google-artifactregistry-auth==1.0.0
language-selector==0.1
launchpadlib==1.10.13
lazr.restfulclient==0.14.2
lazr.uri==1.0.3
lockfile==0.12.2
louis==3.12.0
macaroonbakery==1.3.1
Mako==1.1.0
MarkupSafe==1.1.0
monotonic==1.5
more-itertools==4.2.0
moviepy==1.0.3
mutagen==1.45.1
mysql-connector-python==8.0.27
netifaces==0.10.4
numpy==1.21.2
oauthlib==3.1.1
olefile==0.46
opencv-python==4.5.3.56
paramiko==2.6.0
paypal-checkout-serversdk==1.0.1
paypalhttp==1.0.1
pexpect==4.6.0
phonenumbers==8.12.44
Pillow==8.3.2
platformdirs==2.5.1
pluggy==1.0.0
polib==1.1.1
proglog==0.1.9
protobuf==3.19.1
psycopg2-binary==2.9.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycairo==1.16.2
pycparser==2.20
pycups==1.9.73
PyGObject==3.36.0
PyJWT==2.2.0
pymacaroons==0.13.0
pymediainfo==5.1.0
PyMySQL==1.0.2
PyNaCl==1.3.0
pyOpenSSL==21.0.0
pyRFC3339==1.1
pyrsistent==0.15.5
python-apt==2.0.0+ubuntu0.20.4.7
python-dateutil==2.8.2
python-debian===0.1.36ubuntu1
python-dotenv==0.19.2
python3-openid==3.2.0
PythonVideoConverter==1.0.3
pytz==2021.3
pyxdg==0.26
PyYAML==5.3.1
reportlab==3.5.34
requests==2.26.0
requests-oauthlib==1.3.0
requests-unixsocket==0.2.0
rsa==4.8
s3transfer==0.5.0
SecretStorage==2.3.1
simplejson==3.16.0
six==1.16.0
social-auth-app-django==5.0.0
social-auth-core==4.1.0
sqlparse==0.4.2
systemd-python==234
texttable==1.6.2
tqdm==4.62.3
ubuntu-advantage-tools==27.7
ubuntu-drivers-common==0.0.0
ufw==0.36
unattended-upgrades==0.1
urllib3==1.26.7
usb-creator==0.3.7
virtualenv==20.14.0
wadllib==1.3.3
websocket-client==0.53.0
whitenoise==5.3.0
xkit==0.0.0
zipp==1.0.0
